# Add upsert in `_set_trial_attr_without_commit` for PostgreSQL