--- a/optuna/storages/_rdb/storage.py
+++ b/optuna/storages/_rdb/storage.py
@@ -43,6 +43,7 @@ if TYPE_CHECKING:
     import sqlalchemy
     import sqlalchemy.dialects.mysql as sqlalchemy_dialects_mysql
     import sqlalchemy.dialects.sqlite as sqlalchemy_dialects_sqlite
+    import sqlalchemy.dialects.postgresql as sqlalchemy_dialects_postgresql
     import sqlalchemy.exc as sqlalchemy_exc
     import sqlalchemy.orm as sqlalchemy_orm
     import sqlalchemy.sql.functions as sqlalchemy_sql_functions
@@ -55,6 +56,7 @@ else:
     alembic_script = _LazyImport("alembic.script")
 
     sqlalchemy = _LazyImport("sqlalchemy")
+    sqlalchemy_dialects_postgresql = _LazyImport("sqlalchemy.dialects.postgresql")
     sqlalchemy_dialects_mysql = _LazyImport("sqlalchemy.dialects.mysql")
     sqlalchemy_dialects_sqlite = _LazyImport("sqlalchemy.dialects.sqlite")
     sqlalchemy_exc = _LazyImport("sqlalchemy.exc")
@@ -751,6 +753,15 @@ class RDBStorage(BaseStorage, BaseHeartbeat):
                 set_=dict(value_json=sqlite_insert_stmt.excluded.value_json),
             )
             session.execute(sqlite_upsert_stmt)
+        elif self.engine.name == "postgresql":
+            pg_insert_stmt = sqlalchemy_dialects_postgresql.insert(model_cls).values(
+                trial_id=trial_id, key=key, value_json=json.dumps(value)
+            )
+            pg_upsert_stmt = pg_insert_stmt.on_conflict_do_update(
+                index_elements=[model_cls.trial_id, model_cls.key],
+                set_=dict(value_json=pg_insert_stmt.excluded.value_json),
+            )
+            session.execute(pg_upsert_stmt)
         else:
             # TODO(porink0424): Add support for other databases, e.g., PostgreSQL.
             attribute = model_cls.find_by_trial_and_key(trial, key, session)