@c-bata 2025-09-19
@kAIto47802 Could you review this PR?